requirements.txt
requests
python-dotenv
playwright
fastapi
uvicorn[standard]